pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vdna"
version = "0.1.0"
authors = [
  { name="Benjamin Ramtoula", email="benjamin@robots.ox.ac.uk" },
]
description = "A package to compute and compare Visual DNAs, an approach to represent images."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch",
    "torchvision",
    "huggingface-hub",
    "numpy>=1.24",
    "opencv-python",
    "Pillow",
    "scipy",
    "tqdm>=4.64.1",
]

[project.urls]
"Homepage" = "https://github.com/bramtoula/vdna"
"Bug Tracker" = "https://github.com/bramtoula/vdna/issues"
